requirements.txt
Flask>=3.0,<4.0
Werkzeug>=3.0,<4.0
mysql-connector-python>=9.0
pandas>=2.2,<3.0
numpy>=1.26,<2.0
scikit-learn>=1.4,<2.0
tensorflow==2.16.1
openpyxl>=3.1,<4.0
ollama>=0.3,<1.0
